Rhtslib-package {Rhtslib}	R Documentation

An R packaged HTSlib 1.7

Description

This package is interesting to package developers only, and provides a compiled HTSlib for developers to use in their own packages. See the vignette for instructions on use.

Details

Instructions for use are in the vignette, available with browseVignettes(package="Rhtslib")

Author(s)

Nathaniel Hayden nhayden@fredhutch.org

References

The HTSlib included in the package is based on release 1.7 of HTSlib found in the GitHub repository at https://github.com/samtools/htslib

See Also

pkgconfig

Examples

packageDescription('Rhtslib')
Rhtslib:::htsVersion()

[Package Rhtslib version 1.22.0 Index]